+++ src/services/taptitan/card_function/burst/mod.rs
@@ -45,7 +45,7 @@ pub fn on_proc(
     boss: &mut Boss,
     target_part: BossPartName,
     damage: f64,
-    mirror_force_boost: u32,
+    mirror_force_boost: f64,
     burst_trigger_count: u32,
 ) -> f64 {
     match card.card_id {
